requirements.txt
bcrypt==4.1.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.4
pygame==2.5.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.1.0
Werkzeug==3.0.1
